Move Region to region.py and Bounds to bounds.py
